moving syntax folder to _parser_internals
