+++ setup.py
@@ -40,14 +40,14 @@ MOD_NAMES = [
     "spacy.pipeline.sentencizer",
     "spacy.pipeline.senter",
     "spacy.pipeline.tagger",
-    "spacy.syntax.stateclass",
-    "spacy.syntax._state",
+    "spacy.pipeline._parser_internals.stateclass",
+    "spacy.pipeline._parser_internals._state",
     "spacy.tokenizer",
-    "spacy.syntax.nn_parser",
-    "spacy.syntax._parser_model",
-    "spacy.syntax.nonproj",
-    "spacy.syntax.transition_system",
-    "spacy.syntax.arc_eager",
+    "spacy.pipeline._parser_internals.nn_parser",
+    "spacy.pipeline._parser_internals._parser_model",
+    "spacy.pipeline._parser_internals.nonproj",
+    "spacy.pipeline._parser_internals.transition_system",
+    "spacy.pipeline._parser_internals.arc_eager",
     "spacy.gold.gold_io",
     "spacy.tokens.doc",
     "spacy.tokens.span",
@@ -57,7 +57,7 @@ MOD_NAMES = [
     "spacy.matcher.matcher",
     "spacy.matcher.phrasematcher",
     "spacy.matcher.dependencymatcher",
-    "spacy.syntax.ner",
+    "spacy.pipeline._parser_internals.ner",
     "spacy.symbols",
     "spacy.vectors",
 ]
