+++ spacy/pipeline/_parser_internals/_state.pxd
@@ -4,12 +4,12 @@ from libc.stdint cimport uint32_t, uint64_t
 from cpython.exc cimport PyErr_CheckSignals, PyErr_SetFromErrno
 from murmurhash.mrmr cimport hash64
 
-from ..vocab cimport EMPTY_LEXEME
-from ..structs cimport TokenC, SpanC
-from ..lexeme cimport Lexeme
-from ..symbols cimport punct
-from ..attrs cimport IS_SPACE
-from ..typedefs cimport attr_t
+from ...vocab cimport EMPTY_LEXEME
+from ...structs cimport TokenC, SpanC
+from ...lexeme cimport Lexeme
+from ...symbols cimport punct
+from ...attrs cimport IS_SPACE
+from ...typedefs cimport attr_t
 
 
 cdef inline bint is_space_token(const TokenC* token) nogil:
