+++ spacy/pipeline/_parser_internals/arc_eager.pxd
@@ -1,7 +1,7 @@
 from cymem.cymem cimport Pool
 
 from .stateclass cimport StateClass
-from ..typedefs cimport weight_t, attr_t
+from ...typedefs cimport weight_t, attr_t
 from .transition_system cimport TransitionSystem, Transition
