+++ spacy/pipeline/_parser_internals/ner.pxd
@@ -1,6 +1,6 @@
 from .transition_system cimport TransitionSystem
 from .transition_system cimport Transition
-from ..typedefs cimport attr_t
+from ...typedefs cimport attr_t
 
 
 cdef class BiluoPushDown(TransitionSystem):
